# docs/site/content/en/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Horreum REST API
  description: "Horreum automated change anomaly detection. For more information,\
    \ please see [https://horreum.hyperfoil.io/](https://horreum.hyperfoil.io/)"
  version: 0.10-SNAPSHOT
...
paths:
...
  /api/dataset/bySchema:
    get:
      tags:
      - Dataset
      description: "Retrieve a paginated list of Datasets, with total count, by Schema"
      operationId: listBySchema
      parameters:
      - name: direction
        in: query
        description: Sort direction
        schema:
          $ref: '#/components/schemas/SortDirection'
        example: Ascending
      - name: limit
        in: query
        description: limit the number of results
        schema:
          format: int32
          type: integer
        example: 20
      - name: page
        in: query
        description: filter by page number of a paginated list of Schemas
        schema:
          format: int32
          type: integer
        example: 2
      - name: sort
        in: query
        description: Field name to sort results
        schema:
          default: start
          type: string
        example: name
      - name: uri
        in: query
        description: Schema URI
        required: true
        schema:
          type: string
        example: uri:techempower:0.1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetList'
  /api/dataset/list/{testId}:
    get:
      tags:
      - Dataset
      description: "Retrieve a paginated list of Datasets, with total count, by Test"
      operationId: listByTest
      parameters:
      - name: testId
        in: path
        description: Test ID of test to retrieve list of Datasets
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      - name: direction
        in: query
        description: Sort direction
        schema:
          $ref: '#/components/schemas/SortDirection'
        example: Ascending
      - name: filter
        in: query
        description: JOSN Filter expression to apply to query
        schema:
          type: string
        example:
          buildID: 111111
      - name: limit
        in: query
        description: limit the number of results
        schema:
          format: int32
          type: integer
        example: 20
      - name: page
        in: query
        description: filter by page number of a paginated list of Schemas
        schema:
          format: int32
          type: integer
        example: 2
      - name: sort
        in: query
        description: Field name to sort results
        schema:
          type: string
        example: name
      - name: viewId
        in: query
        description: Optional View ID to filter datasets by view
        schema:
          format: int32
          type: integer
        example: 202
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetList'
  /api/dataset/{datasetId}/labelValues:
    get:
      tags:
      - Dataset
      operationId: labelValues
      parameters:
      - name: datasetId
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LabelValue'
  /api/dataset/{datasetId}/previewLabel:
    post:
      tags:
      - Dataset
      operationId: previewLabel
      parameters:
      - name: datasetId
        in: path
        required: true
        schema:
          format: int32
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Label'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelPreview'
  /api/dataset/{datasetId}/summary:
    get:
      tags:
      - Dataset
      operationId: getSummary
      parameters:
      - name: datasetId
        in: path
        required: true
        schema:
          format: int32
          type: integer
      - name: viewId
        in: query
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetSummary'
  /api/dataset/{id}:
    get:
      tags:
      - Dataset
      description: Retrieve Dataset by ID
      operationId: getDataset
      parameters:
      - name: id
        in: path
        description: Dataset ID to retrieve
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "404":
          description: No Dataset with the given id was found
          content:
            application/json: {}
        "200":
          description: JVM system properties of a particular host.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
...
  /api/run/bySchema:
    get:
      tags:
      - Run
      description: Retrieve a paginated list of Runs with available count for a given
        Schema URI
      operationId: listBySchema
      parameters:
      - name: direction
        in: query
        description: Sort direction
        schema:
          $ref: '#/components/schemas/SortDirection'
        example: Ascending
      - name: limit
        in: query
        description: limit the number of results
        schema:
          format: int32
          type: integer
        example: 20
      - name: page
        in: query
        description: filter by page number of a paginated list of Tests
        schema:
          format: int32
          type: integer
        example: 2
      - name: sort
        in: query
        description: Field name to sort results
        schema:
          type: string
        example: name
      - name: uri
        in: query
        description: Schema URI
        required: true
        schema:
          type: string
        example: uri:my-schema:0.1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunsSummary'
...
  /api/test:
    get:
      tags:
      - Test
      description: Retrieve a paginated list of Tests with available count
      operationId: list
      parameters:
      - name: direction
        in: query
        description: Sort direction
        schema:
          $ref: '#/components/schemas/SortDirection'
        example: Ascending
      - name: limit
        in: query
        description: limit the number of results
        schema:
          format: int32
          type: integer
        example: 20
      - name: page
        in: query
        description: filter by page number of a paginated list of Tests
        schema:
          format: int32
          type: integer
        example: 2
      - name: roles
        in: query
        description: "__my, __all or a comma delimited  list of roles"
        schema:
          type: string
        example: __my
      - name: sort
        in: query
        description: Field name to sort results
        schema:
          default: name
          type: string
        example: name
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestQueryResult'
    post:
      tags:
      - Test
      description: Create a new test
      operationId: add
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Test'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
...
  /api/test/summary:
    get:
      tags:
      - Test
      description: Retrieve a summary of Tests in a folder
      operationId: summary
      parameters:
      - name: folder
        in: query
        description: name of the Folder containing the Tests
        schema:
          type: string
        example: My Team Folder
      - name: roles
        in: query
        description: "\"__my\", \"__all\" or a comma delimited  list of roles"
        schema:
          type: string
        example: __my
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestListing'
...
components:
  schemas:
    Access:
      description: Access rights for the test. This defines the visibility of the
        Test in the UI
      enum:
      - 0
      - 1
      - 2
      type: integer
      example: 0
      nullable: false
...
    Dataset:
      description: A dataset is the JSON document used as the basis for all comparisons
        and reporting
      required:
      - start
      - stop
      - testid
      - owner
      - access
      - data
      - ordinal
      type: object
      properties:
        id:
          format: int32
          description: Dataset Unique ID
          type: integer
          example: 101
        start:
          format: date-time
          description: Dataset Start timestamp
          type: string
          example: "1698013206000"
          nullable: false
        stop:
          format: date-time
          description: Dataset Stop timestamp
          type: string
          example: "1698013206000"
          nullable: false
        description:
          description: Run description
          type: string
          example: Run on AWS with m7g.large
        testid:
          format: int32
          description: Test ID that Dataset relates to
          type: integer
          example: 101
          nullable: false
        owner:
          description: Name of the team that owns the test. Users must belong to the
            team that owns a test to make modifications
          type: string
          example: performance-team
          nullable: false
        access:
          description: Access rights for the test. This defines the visibility of
            the Test in the UI
          enum:
          - 0
          - 1
          - 2
          type: integer
          example: 0
          nullable: false
        data:
          description: Data payload
          type: string
          nullable: false
        ordinal:
          format: int32
          description: Dataset ordinal for ordered list of Datasets derived from a
            Run
          type: integer
          example: 1
          nullable: false
        validationErrors:
          description: Collection of Validation Errors
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        runId:
          format: int32
          description: Run ID that Dataset relates to
          type: integer
          example: 101
...
    DatasetList:
      description: Result containing a subset of Dataset Summaries and the total count
        of available. Used in paginated tables
      required:
      - total
      - datasets
      type: object
      properties:
        total:
          format: int64
          description: Total number of Dataset Summaries available
          type: integer
          example: 64
        datasets:
          description: List of Dataset Summaries. This is often a subset of total
            available.
          type: array
          items:
            $ref: '#/components/schemas/DatasetSummary'
          nullable: false
...
    DatasetSummary:
      required:
      - id
      - runId
      - ordinal
      - testId
      - testname
      - start
      - stop
      - owner
      - access
      - schemas
      type: object
      properties:
        id:
          format: int32
          description: Unique Dataset ID
          type: integer
          example: 101
        runId:
          format: int32
          description: Run ID that Dataset relates to
          type: integer
          example: 202
        ordinal:
          format: int32
          description: Ordinal position of Dataset Summary on returned List
          type: integer
          example: 3
        testId:
          format: int32
          description: Test ID that Dataset relates to
          type: integer
          example: 202
        testname:
          description: Test name that the Dataset relates to
          type: string
          example: my-comprehensive-benchmark
          nullable: false
        description:
          description: Dataset description
          type: string
          example: Run on AWS with m7g.large
        start:
          format: int64
          description: Dataset Start timestamp
          type: integer
          example: 1698013206000
        stop:
          format: int64
          description: Dataset Stop timestamp
          type: integer
          example: 1698013206000
        owner:
          description: Name of the team that owns the test. Users must belong to the
            team that owns a test to make modifications
          type: string
          example: performance-team
        access:
          $ref: '#/components/schemas/Access'
        view:
          description: View definition
          type: string
        schemas:
          description: List of Schema usages
          type: array
          items:
            $ref: '#/components/schemas/SchemaUsage'
        validationErrors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    Extractor:
      description: "An Extractor defines how values are extracted from a JSON document,\
        \ for use in Labels etc."
      required:
      - name
      - jsonpath
      - array
      type: object
      properties:
        name:
          description: Name of extractor. This name is used in Combination Functions
            to refer to values by name
          type: string
          example: buildID
          nullable: false
        jsonpath:
          description: JSON path expression defining the location of the extractor
            value in the JSON document. This is a pSQL json path expression
          type: string
          example: $.buildInfo.buildID
          nullable: false
        array:
          description: Does the JSON path expression reference an Array?
          type: boolean
          example: false
          nullable: false
...
    Label:
      description: "A Label is a core component of Horreum, defining which components\
        \ of the JSON document are part of a KPI and how the metric values are calculated"
      required:
      - id
      - name
      - extractors
      - filtering
      - metrics
      - owner
      - access
      - schemaId
      type: object
      properties:
        id:
          format: int32
          description: Unique ID for Label
          type: integer
          example: 101
        name:
          description: "Name for label. NOTE: all Labels are considered to have the\
            \ same semantic meaning throughout the entire system"
          type: string
          example: Throughput
          nullable: false
        extractors:
          description: "A collection of Extractors, that will be combined in the Combination\
            \ Function"
          type: array
          items:
            $ref: '#/components/schemas/Extractor'
          nullable: false
        function:
          description: A Combination Function that defines how values from Extractors
            are combined to produce a Label Value
          type: string
          example: "value => { return ((value.reduce((a,b) => a+b))/value.length*1000).toFixed(3);\
            \ }"
        filtering:
          description: Is Label a filtering label? Filtering labels contains values
            that are used to filter datasets for comparison
          type: boolean
          example: true
          nullable: false
        metrics:
          description: Is Label a metrics label? Metrics labels are contain Metrics
            that are used for comparison
          type: boolean
          example: true
          nullable: false
        owner:
          description: Name of the team that owns the test. Users must belong to the
            team that owns a test to make modifications
          type: string
          example: performance-team
        access:
          description: Access rights for the test. This defines the visibility of
            the Test in the UI
          enum:
          - 0
          - 1
          - 2
          type: integer
          example: 0
          nullable: false
        schemaId:
          format: int32
          description: Schema ID that the Label relates to
          type: integer
          example: 101
          nullable: false
...
    LabelPreview:
      description: Preview a Label Value derived from a Dataset Data. A preview allows
        users to apply a Label to a dataset and preview the Label Value result and
        processing errors in the UI
      type: object
      properties:
        value:
          description: "Value value extracted from Dataset. This can be a scalar,\
            \ array or JSON object"
          type: string
        output:
          description: Description of errors occurred attempting to generate Label
            Value Preview
          type: string
    LabelValue:
      description: Label Value derived from Label definition and Dataset Data
      required:
      - id
      - name
      - schema
      type: object
      properties:
        id:
          format: int32
          description: Unique ID for Label Value
          type: integer
          example: 101
        name:
          description: Label name
          type: string
          example: buildID
          nullable: false
        schema:
          allOf:
          - $ref: '#/components/schemas/SchemaDescriptor'
          - description: Summary description of Schema
            nullable: false
        value:
          description: "Value value extracted from Dataset. This can be a scalar,\
            \ array or JSON object"
          type: string
          example: "1724"
...
    RunsSummary:
      required:
      - total
      - runs
      type: object
      properties:
        total:
          format: int64
          description: Total count of Runs visible
          type: integer
          example: 1
        runs:
          description: List of Run Summaries
          type: array
          items:
            $ref: '#/components/schemas/RunSummary'
          nullable: false
...
    SchemaDescriptor:
      required:
      - id
      - name
      - uri
      type: object
      properties:
        id:
          format: int32
          description: Schema unique ID
          type: integer
          example: 1
        name:
          description: Schema name
          type: string
          example: my-benchmark-schema
          nullable: false
        uri:
          description: Schema name
          type: string
          example: uri:my-schmea:0.1
          nullable: false
...
    SchemaUsage:
      required:
      - id
      - name
      - uri
      - source
      - type
      - hasJsonSchema
      type: object
      properties:
        id:
          format: int32
          description: Schema unique ID
          type: integer
          example: 1
        name:
          description: Schema name
          type: string
          example: my-benchmark-schema
          nullable: false
        uri:
          description: Schema name
          type: string
          example: uri:my-schmea:0.1
          nullable: false
        source:
          format: int32
          description: "Source of schema usage, 0 is data, 1 is metadata. DataSets\
            \ always use 0"
          type: integer
          example: 1
        type:
          format: int32
          description: "Location of Schema Usage, 0 for Run, 1 for Dataset"
          type: integer
          example: 1
        key:
          description: Ordinal position of schema usage in Run/Dataset
          type: string
          example: "1"
        hasJsonSchema:
          description: Does schema have a JSON validation schema defined?
          type: boolean
          example: false
    SortDirection:
      enum:
      - Ascending
      - Descending
      type: string
...
    Test:
      description: Represents a Test. Tests are typically equivalent to a particular
        benchmark
      required:
      - id
      - name
      - owner
      - access
      - notificationsEnabled
      type: object
      properties:
        id:
          format: int32
          description: Unique Test id
          type: integer
          example: 101
        name:
          description: Test name
          type: string
          example: my-comprehensive-benchmark
          nullable: false
        folder:
          description: Name of folder that the test is stored in. Folders allow tests
            to be organised in the UI
          type: string
          example: My Team Folder
        description:
          description: Description of the test
          type: string
          example: Comprehensive benchmark to tests the limits of any system it is
            run against
        owner:
          description: Name of the team that owns the test. Users must belong to the
            team that owns a test to make modifications
          type: string
          example: performance-team
          nullable: false
        access:
          description: Access rights for the test. This defines the visibility of
            the Test in the UI
          enum:
          - 0
          - 1
          - 2
          type: integer
          example: 0
          nullable: false
        tokens:
          description: Array of API tokens associated with test
          type: array
          items:
            $ref: '#/components/schemas/TestToken'
        timelineLabels:
          description: List of label names that are used for determining metric to
            use as the time series
          type: array
          items:
            type: string
          example:
          - timestamp
        timelineFunction:
          description: Label function to modify timeline labels to a produce a value
            used for ordering datapoints
          type: string
          example: timestamp => timestamp
        fingerprintLabels:
          description: 'Array of Label names that are used to create a fingerprint '
          type: array
          items:
            type: string
          example:
          - build_tag
        fingerprintFilter:
          description: Filter function to filter out datasets that are comparable
            for the purpose of change detection
          type: string
          example: value => value === "true"
        compareUrl:
          description: URL to external service that can be called to compare runs.  This
            is typically an external reporting/visulization service
          type: string
          example: "(ids, token) => 'http://repoting.example.com/report/specj?q='\
            \ + ids.join('&q=') + \"&token=\"+token"
        transformers:
          description: Array for transformers defined for the Test
          type: array
          items:
            $ref: '#/components/schemas/Transformer'
        notificationsEnabled:
          description: Are notifications enabled for the test
          type: boolean
          example: true
          nullable: false
    TestListing:
      type: object
      properties:
        tests:
          description: Array of Test Summaries
          type: array
          items:
            $ref: '#/components/schemas/TestSummary'
    TestQueryResult:
      required:
      - tests
      - count
      type: object
      properties:
        tests:
          description: Array of Tests
          type: array
          items:
            $ref: '#/components/schemas/Test'
          nullable: false
        count:
          format: int64
          description: Count of available tests. This is a count of tests that the
            current user has access to
          type: integer
          example: 64
...